+++ Blurred-ios/FollowListViewController.swift
@@ -20,6 +20,7 @@ class FollowListViewController: UIViewController, UITableViewDataSource {
     override func viewDidLoad() {
         super.viewDidLoad()
         downloadJson()
+        tableView.tableFooterView = UIView()
         // Do any additional setup after loading the view.
     }
     func downloadJson() { // Still not done we need to add the user's butt image
@@ -28,6 +29,9 @@ class FollowListViewController: UIViewController, UITableViewDataSource {
         guard let downloadURL = url else { return }
         URLSession.shared.dataTask(with: downloadURL) { (data, urlResponse, error) in
             guard let data = data, error == nil, urlResponse != nil else {
+                DispatchQueue.main.async {
+                    self.showNoResponseFromServer()
+                }
                 print("uh oh")
                 return
             }
@@ -75,24 +79,30 @@ class FollowListViewController: UIViewController, UITableViewDataSource {
         request.httpMethod = "GET"
         let task = URLSession.shared.dataTask(with: request) { (data: Data?, response: URLResponse?, error: Error?) in
             if error != nil {
-                self.showErrorContactingServer()
+                DispatchQueue.main.async {
+                    self.showErrorContactingServer()
+                }
                 return
             }
             do {
                 let json = try JSONSerialization.jsonObject(with: data!, options: .mutableContainers) as? NSDictionary
                 if let parseJSON = json {
                     let imageUrl: String? = parseJSON["image_url"] as? String
-                    let railsUrl = URL(string: "http://10.0.0.2:3000\(imageUrl ?? "/public/default-avatar-3")")
+                    let railsUrl = URL(string: "http://10.0.0.2:3000\(imageUrl ?? "/assets/fallback/default-avatar-3.png")")
                     DispatchQueue.main.async {
                         Nuke.loadImage(with: railsUrl!, into: cell.followingAvatar)
-                        }
+                    }
                 } else {
-                    self.showErrorContactingServer()
+                    DispatchQueue.main.async {
+                        self.showErrorContactingServer()
+                    }
                     print(error ?? "No error")
                 }
             } catch {
+                DispatchQueue.main.async {
                     self.showNoResponseFromServer()
-                    print(error)
+                }
+                print(error)
                 }
         }
         task.resume()